Move filesystem logic to models
